refactor: [WC-52] NoCardGuide ts전환& Text stories파일 ts전환
