+++ src/components/molecule/NoCardGuide/NoCardGuide.stories.tsx
@@ -1,0 +1,10 @@
+import { NoCardGuide } from '@/components';
+
+export default {
+  title: 'Component/molecule/NoCardGuide',
+  component: NoCardGuide,
+};
+
+export const Default = () => {
+  return <NoCardGuide />;
+};
