+++ src/components/molecule/NoCardGuide/NoCardGuide.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
+import type { HTMLAttributes, ReactElement } from 'react';
 import styled from '@emotion/styled';
 import { Text } from '@/components';
 import Common from '@/styles';
 
-const NoCardGuide = ({ ...props }) => {
+const NoCardGuide = ({
+  ...props
+}: HTMLAttributes<HTMLDivElement>): ReactElement => {
   return (
     <TextContainer {...props}>
       <StyledText>카드 목록이 없습니다</StyledText>
